+++ expands/install.sh
@@ -6,7 +6,7 @@
 # See /LICENSE for more information.
 #
 
-install_addon() {
+if [ "${1}" = "late" ]; then
   echo "Installing addon expands - ${1}"
   mkdir -p "/tmpRoot/usr/arc/addons/"
   cp -pf "${0}" "/tmpRoot/usr/arc/addons/"
@@ -14,7 +14,6 @@ install_addon() {
   cp -pf /usr/bin/expands.sh /tmpRoot/usr/bin/expands.sh
 
   mkdir -p "/tmpRoot/usr/lib/systemd/system"
-  DEST="/tmpRoot/usr/lib/systemd/system/expands.service"
   {
     echo "[Unit]"
     echo "Description=Expanded miscellaneous"
@@ -27,12 +26,11 @@ install_addon() {
     echo
     echo "[Install]"
     echo "WantedBy=multi-user.target"
-  } >"${DEST}"
+  } >"/tmpRoot/usr/lib/systemd/system/expands.service"
+
   mkdir -p /tmpRoot/usr/lib/systemd/system/multi-user.target.wants
   ln -vsf /usr/lib/systemd/system/expands.service /tmpRoot/usr/lib/systemd/system/multi-user.target.wants/expands.service
-}
-
-uninstall_addon() {
+elif [ "${1}" = "uninstall" ]; then
   echo "Uninstalling addon expands - ${1}"
 
   rm -f "/tmpRoot/usr/lib/systemd/system/multi-user.target.wants/expands.service"
@@ -41,13 +39,4 @@ uninstall_addon() {
 [ ! -f "/tmpRoot/usr/arc/revert.sh" ] && echo '#!/usr/bin/env bash' >/tmpRoot/usr/arc/revert.sh && chmod +x /tmpRoot/usr/arc/revert.sh
   echo "/usr/bin/expands.sh -r" >>/tmpRoot/usr/arc/revert.sh
   echo "rm -f /usr/bin/expands.sh" >>/tmpRoot/usr/arc/revert.sh
-}
-
-case "${1}" in
-  late)
-    install_addon "${1}"
-    ;;
-  uninstall)
-    uninstall_addon "${1}"
-    ;;
-esac
+fi
